properly handling when a module cannot init!
